move AnalysisScope to cudaq-common
