+++ runtime/common/AnalysisScope.h
