+++ runtime/common/CMakeLists.txt
@@ -11,6 +11,7 @@ set(LIBRARY_NAME cudaq-common)
 
 set(COMMON_EXTRA_DEPS "")
 set(COMMON_RUNTIME_SRC
+  AnalysisScope.cpp
   CodeGenConfig.cpp
   cudaq_json.cpp
   CustomOp.cpp
