+++ runtime/nvqir/CMakeLists.txt
@@ -11,7 +11,6 @@ set(LIBRARY_NAME nvqir)
 set (CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wno-ctad-maybe-unsupported")
 
 set(NVQIR_RUNTIME_SRC
-  AnalysisScope.cpp
   QIRTypes.cpp
   NVQIR.cpp
   dem/DemScope.cpp
